+++ src/Type/RichTextItalic.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Phptg\BotApi\Type;
+
+use Phptg\BotApi\ParseResult\ValueProcessor\RichTextValue;
+
+/**
+ * @see https://core.telegram.org/bots/api#richtextitalic
+ *
+ * @api
+ */
+final readonly class RichTextItalic implements RichText
+{
+    public function __construct(
+        #[RichTextValue]
+        public string|array|RichText $text,
+    ) {}
+
+    public function getType(): string
+    {
+        return 'italic';
+    }
+}